Extract CommitDiff user control
