+++ GitUI/HelperDialogs/FormCommitDiff.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-using GitCommands;
+﻿using System;
 
 namespace GitUI.HelperDialogs
 {
@@ -12,9 +9,6 @@ namespace GitUI.HelperDialogs
         {
             InitializeComponent();
             Translate();
-            DiffText.ExtraDiffArgumentsChanged += DiffText_ExtraDiffArgumentsChanged;
-            DiffFiles.Focus();
-            DiffFiles.SetDiffs();
         }
 
         private FormCommitDiff()
@@ -25,49 +19,13 @@ namespace GitUI.HelperDialogs
         public FormCommitDiff(GitUICommands commands, string revisionGuid)
             : this(commands)
         {
-            // We cannot use the GitRevision from revision grid. When a filtered commit list
-            // is shown (file history/normal filter) the parent guids are not the 'real' parents,
-            // but the parents in the filtered list.
-            GitRevision revision = Module.GetRevision(revisionGuid);
-
-            if (revision != null)
-            {
-                commitInfo.Revision = revision;
-
-                DiffFiles.SetDiffs(new[] { revision });
-
-                Text = "Diff - " + GitRevision.ToShortSha(revision.Guid) + " - " + revision.AuthorDate + " - " + revision.Author + " - " + Module.WorkingDir;
-            }
+            CommitDiff.TextChanged += CommitDiff_TextChanged;
+            CommitDiff.SetRevision(revisionGuid);
         }
 
-        private async void DiffFiles_SelectedIndexChanged(object sender, EventArgs e)
+        private void CommitDiff_TextChanged(object sender, EventArgs e)
         {
-            try
-            {
-                await ViewSelectedDiff();
-            }
-            catch (OperationCanceledException)
-            {
-            }
-        }
-
-        private async Task ViewSelectedDiff()
-        {
-            if (DiffFiles.SelectedItem != null && DiffFiles.Revision != null)
-            {
-                await DiffText.ViewChanges(DiffFiles.SelectedItemParent?.Guid, DiffFiles.Revision?.Guid, DiffFiles.SelectedItem, string.Empty);
-            }
-        }
-
-        private async void DiffText_ExtraDiffArgumentsChanged(object sender, EventArgs e)
-        {
-            try
-            {
-                await ViewSelectedDiff();
-            }
-            catch (OperationCanceledException)
-            {
-            }
+            Text = CommitDiff.Text;
         }
     }
 }
